backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
pillow==11.3.0
psycopg==3.2.9
sqlparse==0.5.3
tzdata==2025.2